+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - File Area: Add Observed/Reference Files
 - File Area: Add support for .pdf files
 - Settings: Add Settings - Figures - Word Clouds - Font Path
+- Settings: Add Settings - Measures - Bayes Factor
 - Settings: Add Settings - Measures - Statistical Significance - Log-likelihood Ratio Test / Welch's t-test / z-score (Berry-Rogghe)
 - Settings: Add Settings - Measures - Statistical Significance - Student's t-test (1-sample) / Student's t-test (2-sample) / Welch's t-test - Direction
 - Settings: Add Settings - Tables - Profiler